+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased - 2021-09-08
+
+### Changed
+
+- Optimize db updates with updateDocuments(). Total db operations are now 2 + <number of scenes>.
+
 ## [0.4.0] - 2021-09-07
 
 ### Added