+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## Unreleased - 2021-09-08
+## [0.5.0] - 2021-09-09
+
+### Added
+
+- Additional docs on defaults and settings to convert to local data images
 
 ### Changed
 
@@ -40,6 +44,7 @@
 
 - Add Update 5eTools Images macro with UI
 
+[0.5.0]: https://github.com/surged20/foundryvtt-update-image-macro/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/surged20/foundryvtt-update-image-macro/compare/0.3.1...0.4.0
 [0.3.1]: https://github.com/surged20/foundryvtt-update-image-macro/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/surged20/foundryvtt-update-image-macro/compare/0.2.1...0.3.0